update code: interface 90%

+++ ant-api/src/main/java/com/bx/ant/service/DeliverOrderShopServiceI.java
@@ -127,5 +127,5 @@ public interface DeliverOrderShopServiceI {
 	DeliverOrderShop editStatusByHql(DeliverOrderShop deliverOrderShop,String status,String shopPayStatus);
 
 
-    DeliverOrderShopView getView(Integer id);
+    DeliverOrderShopView getView(Long id);
 }
